"mm: swap: don't delay swap free for fast swap devices" There are couple of issues with swapcache usage when ZRAM is used
as swap device.

1) Kernel does a swap readahead which can be around 6 to 8 pages

depending on total ram, which is not required for zram since

accesses are fast.

2) Kernel delays the freeing up of swapcache expecting a later hit,

which again is useless in the case of zram.

3) This is not related to swapcache, but zram usage itself.

As mentioned in (2) kernel delays freeing of swapcache, but along with

that it delays zram compressed page free also. i.e. there can be 2 copies,

though one is compressed.

This patch addresses these issues using two new flags

QUEUE_FLAG_FAST and SWP_FAST, to indicate that accesses to the device

will be fast and cheap, and instructs the swap layer to free up

swap space agressively, and not to do read ahead.

Change-Id: I5d2d5176a5f9420300bb2f843f6ecbdb25ea80e4

Signed-off-by: Vinayak Menon <vinmenon@codeaurora.org>