Add SkyDragon Optimization for msm8994
Full set of compiler flags for msm8994 based on optimizing at -O2 or -O3 at most.

Graphite available when toolchain supports it, disabled by default.

Not aligning generally hurts performance on this chip, so we align to keep it regardless of overall -O level.

Hoist and loop pressure slow the compiler/build process but generate smaller, faster code which is what we want. So we turn this on.

Global common subexpression elimination can definitely speed up code by eliminating things we don't need or pulling them out of the loop for easier access. We want all of this turned on.

We enable schedule pressure and speculative load to further decrease code size and speed up code.

We compile for cortex-a57.cortex-a53 with all features turned on, as this is what the msm8994/snapdragon810 is. We make sure the compiler knows it has a 64-byte cache line size, a 32-byte l1 cache size, and 2MB l2-cache so we generate the best code for this specific configuration.

Disable debugging. Keep everything in order.
Make sure to keep dropping the frame pointer.
Enable induction variable optimization on all trees.
Enable swing modulo scheduling to further optimize loops.